extract corruption checker class in separate header file
